agent-backend/requirements.txt

azure-identity
azure-keyvault-secrets
openai
requests
langchain
